Cover Letter generator/requirements.txt
openai
langchain
streamlit
streamlit-chat
tiktoken
python-dotenv
streamlit-authenticator